Promote 'alpha' as a default command
Change 'run' to 'monolithic'
