+++ engine/cmd/root.go
@@ -10,7 +10,7 @@ var rootCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(runCmd())
-	rootCmd.AddCommand(alphaCmd())
+	rootCmd.AddCommand(runMonoCmd())
 	rootCmd.AddCommand(pullCmd())
 	rootCmd.SilenceUsage = true
 }
